+++ back/babble/src/main/java/gg/babble/babble/util/UrlPhaser.java
@@ -1,4 +1,4 @@
-package gg.babble.babble.config;
+package gg.babble.babble.util;
 
 import java.util.Map;
 import org.springframework.util.AntPathMatcher;
@@ -7,6 +7,7 @@ public enum UrlPhaser {
     ROOM_SUBSCRIBE_URL_PATTERN("/topic/rooms/{roomId}/*");
 
     public static final AntPathMatcher PATH_MATCHER = new AntPathMatcher();
+
     private final String urlPattern;
 
     UrlPhaser(final String urlPattern) {
