+++ back/babble/src/test/java/gg/babble/babble/config/UrlPhaserTest.java
@@ -3,6 +3,7 @@ package gg.babble.babble.config;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import gg.babble.babble.util.UrlPhaser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
